sync origin

+++ backend/src/main/java/io/metersphere/controller/ProjectController.java
@@ -9,6 +9,7 @@ import io.metersphere.commons.utils.Pager;
 import io.metersphere.commons.utils.SessionUtils;
 import io.metersphere.controller.request.ProjectRequest;
 import io.metersphere.dto.ProjectDTO;
+import io.metersphere.service.CheckPermissionService;
 import io.metersphere.service.ProjectService;
 import org.apache.shiro.authz.annotation.Logical;
 import org.apache.shiro.authz.annotation.RequiresRoles;
@@ -22,6 +23,8 @@ import java.util.List;
 public class ProjectController {
     @Resource
     private ProjectService projectService;
+    @Resource
+    private CheckPermissionService checkPermissionService;
 
     @GetMapping("/listAll")
     public List<ProjectDTO> listAll() {
@@ -71,6 +74,7 @@ public class ProjectController {
     @GetMapping("/delete/{projectId}")
     @RequiresRoles(value = {RoleConstants.TEST_MANAGER, RoleConstants.TEST_USER,}, logical = Logical.OR)
     public void deleteProject(@PathVariable(value = "projectId") String projectId) {
+        checkPermissionService.checkProjectOwner(projectId);
         projectService.deleteProject(projectId);
     }
 
@@ -79,4 +83,5 @@ public class ProjectController {
     public void updateProject(@RequestBody Project Project) {
         projectService.updateProject(Project);
     }
+
 }
